perf: SWR string-cache for /swap/pools, /tickers, /pools/:id/ticks
route-cache caches the raw object and re-runs JSON.stringify on every hit,
so cacheSeconds removed Mongo load but not CPU — the 10MB pools payload was
serialized once per request. Extract the analytics SWR core into a shared
server/services/swrCache.ts and add getSwrString, which caches the already
serialized body and serves it raw (zero parse, zero stringify on the hot
path) with stale-while-revalidate background refresh.

- swrCache.ts: shared getSwrPayload (used by analytics) + getSwrString
- analytics.ts: import the shared core, drop the duplicated ~170 lines
- swap.ts /pools, /pools/:id/ticks and spot.js /tickers: getSwrString
